refactor(cef): drop the cef://localhost scheme
Nothing has navigated to `cef://localhost/` since the app moved to its own
`vmux://` embedded scheme. The three constructors that produced such URLs —
`WebviewSource::{local, embedded, inline}` — had no callers, and neither did
`custom_scheme_{host_root, host_url, document_url}`; every webview in the app
is spawned with `WebviewSource::new`.

With `inline` gone the `InlineHtml` variant is unreachable, so `WebviewSource`
collapses to the tuple struct it always was, taking the inline HTML store, its
id counter, its despawn hook and the `__inline__/` URI parsing with it. The
scheme itself is no longer registered: two `add_custom_scheme` calls and three
`register_scheme_handler_factory` calls are removed, along with the resolver
branch that mapped those URLs onto asset paths.

The module was named after the scheme, not the job. It serves `vmux://`,
`file://` and the vault origin, so it is `custom_scheme` now.

Also removes an `EnvGuard` orphaned by the source-scan test deletion earlier in
this branch, which `-D warnings` would have failed on.
