+++ patches/bevy_cef-0.5.2/src/common/custom_scheme.rs
@@ -3,12 +3,12 @@ use bevy::prelude::*;
 mod asset_loader;
 pub(crate) mod responser;
 
-use crate::common::localhost::asset_loader::LocalSchemeAssetLoaderPlugin;
+use crate::common::custom_scheme::asset_loader::LocalSchemeAssetLoaderPlugin;
 
 /// A plugin that adds support for handling local scheme requests in Bevy applications.
-pub(crate) struct LocalHostPlugin;
+pub(crate) struct CustomSchemePlugin;
 
-impl Plugin for LocalHostPlugin {
+impl Plugin for CustomSchemePlugin {
     fn build(&self, app: &mut App) {
         app.add_plugins((responser::ResponserPlugin, LocalSchemeAssetLoaderPlugin));
     }
