+++ patches/bevy_cef-0.5.2/src/common/custom_scheme/asset_loader.rs
