+++ patches/bevy_cef_core-0.5.2/src/browser_process.rs
@@ -4,13 +4,13 @@ mod browsers;
 mod client_handler;
 mod command_line_config;
 mod context_menu_handler;
+mod custom_scheme;
 mod display_handler;
 mod extensions;
 #[cfg(target_os = "macos")]
 mod keepalive;
 mod life_span_handler;
 mod load_handler;
-mod localhost;
 mod message_pump;
 mod permission_handler;
 mod renderer_handler;
@@ -23,13 +23,13 @@ pub use browsers::*;
 pub use client_handler::*;
 pub use command_line_config::*;
 pub use context_menu_handler::*;
+pub use custom_scheme::*;
 pub use display_handler::*;
 pub use extensions::*;
 #[cfg(target_os = "macos")]
 pub use keepalive::*;
 pub use life_span_handler::*;
 pub use load_handler::*;
-pub use localhost::*;
 pub use message_pump::*;
 pub use permission_handler::*;
 pub use renderer_handler::*;
