+++ patches/bevy_cef_core-0.5.2/src/browser_process/app.rs
@@ -4,7 +4,7 @@ use crate::browser_process::MessageLoopTimer;
 use crate::browser_process::MessageLoopWakePolicy;
 use crate::browser_process::browser_process_handler::BrowserProcessHandlerBuilder;
 use crate::browser_process::browser_process_handler::WakeProxy;
-use crate::util::{SCHEME_CEF, cef_scheme_flags};
+use crate::util::cef_scheme_flags;
 use cef::rc::{Rc, RcImpl};
 use cef::{
     BrowserProcessHandler, CefString, CommandLine, ImplApp, ImplCommandLine, ImplSchemeRegistrar,
@@ -121,7 +121,6 @@ impl ImplApp for BrowserProcessAppBuilder {
 
     fn on_register_custom_schemes(&self, registrar: Option<&mut SchemeRegistrar>) {
         if let Some(registrar) = registrar {
-            registrar.add_custom_scheme(Some(&SCHEME_CEF.into()), cef_scheme_flags() as _);
             let cfg = crate::util::resolved_cef_embedded_page_config();
             registrar.add_custom_scheme(Some(&cfg.scheme.as_str().into()), cef_scheme_flags() as _);
         }
