+++ patches/bevy_cef_core-0.5.2/src/browser_process/custom_scheme/data_responser.rs
